pyproject.toml
[tool.poetry]
name = "possebles"
version = "0.1.0"
description = ""
authors = ["Pierre-Yves Gillier <github@pygillier.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
sqlalchemy = "^1.4.46"
alembic = "^1.9.2"
loguru = "^0.6.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
feedparser = "^6.0.10"
mastodon-py = "^1.8.0"



[tool.poetry.group.dev.dependencies]
openapi-generator-cli = "^4.3.1"
invoke = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
